+++ lib/router/index/index_widget.dart
@@ -17,6 +17,7 @@ import 'package:nostrmo/router/index/index_pc_drawer_wrapper.dart';
 import 'package:provider/provider.dart';
 
 import '../../features/communities/communities_screen.dart';
+import '../../features/create_community/create_community_dialog.dart';
 import '../../generated/l10n.dart';
 import '../../main.dart';
 import '../../provider/index_provider.dart';
@@ -24,7 +25,6 @@ import '../../provider/settings_provider.dart';
 import '../../util/auth_util.dart';
 import '../../util/table_mode_util.dart';
 import '../dm/dm_widget.dart';
-import '../group/create_community_dialog.dart';
 import '../login/login_widget.dart';
 import '../search/search_widget.dart';
 import 'index_app_bar.dart';
